Unify getParams actions across files
