+++ src/modules/auth/containers/LoginScreen.js
@@ -5,6 +5,7 @@ import { Container, Row, Col } from 'reactstrap';
 
 import LoginCard from './LoginCard.js';
 import RegistrationCard from './RegistrationCard.js';
+import { getParams } from '../utils/paramsActions';
 
 function LoginScreen(props) {
 	const { t } = useTranslation();
@@ -92,17 +93,6 @@ function LoginScreen(props) {
 		} catch (e) {
 			console.error("Failed to fetch external login services", e);
 		}
-	}
-
-	function getParams(param) {
-		let parameter = undefined;
-		const i = window.location.hash.indexOf('?');
-		if (i > -1) {
-			const qs = window.location.hash.substring(i+1);
-			const params = new URLSearchParams(qs);
-			parameter = params.get(param);
-		}
-		return parameter;
 	}
 
 	return (
